requirements.txt
pymupdf>=1.24.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
numpy>=1.26.0
tqdm>=4.66.0

fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.1

google-genai>=0.1.0

